requirements.txt
aws-cdk-lib==2.93.0
constructs==10.2.70